Rework: Split into 'Core' and 'Rig' crates with Absolute/Relative resolvers.
