+++ .github/workflows/rust.yml
@@ -46,17 +46,19 @@ jobs:
         working-directory: src
         shell: bash
         run: |
-          SEARCH_RESULT=$(cargo search "^rig-coding-tools$" --limit 1)
+          # Note: binstall is available after devops-rust-test-and-coverage@v1 call
+          cargo +stable binstall --no-confirm cargo-semver-checks --force
+          rustup +stable target add ${{ matrix.target }}
 
-          if echo "$SEARCH_RESULT" | grep -q "^rig-coding-tools "; then
-              # Run semver checks on stable, because nightly sometimes gets borked in cargo-semver-checks.
-              rustup +stable target add ${{ matrix.target }}
-              # Note: binstall is available after devops-rust-test-and-coverage@v1 call
-              cargo +stable binstall --no-confirm cargo-semver-checks --force
-              cargo +stable semver-checks --target ${{ matrix.target }} 
-          else
-              echo "No previous version found on crates.io. Skipping semver checks."
-          fi
+          for CRATE in "coding-tools-core" "coding-tools-rig"; do
+            SEARCH_RESULT=$(cargo search "^${CRATE}$" --limit 1)
+            if echo "$SEARCH_RESULT" | grep -q "^${CRATE} "; then
+              echo "Running semver checks for ${CRATE}..."
+              cargo +stable semver-checks -p "${CRATE}" --target ${{ matrix.target }}
+            else
+              echo "No previous version of ${CRATE} found on crates.io. Skipping semver checks."
+            fi
+          done
 
       - name: Check documentation is valid
         if: github.event_name == 'pull_request' || startsWith(github.ref, 'refs/tags/')
@@ -89,7 +91,9 @@ jobs:
         uses: Reloaded-Project/devops-publish-action@v3
         with:
           rust-crates-io-token: ${{ secrets.CRATES_IO_TOKEN }}
-          rust-cargo-project-paths: src/rig-coding-tools
+          rust-cargo-project-paths: |
+            src/coding-tools-core
+            src/coding-tools-rig
           compression-tool: 7z
           artifact-groups-file: .github/artifact-groups.yml
           changelog-enabled: "true"
