+++ src/coding-tools-rig/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "coding-tools-rig"
+version = "0.1.0"
+edition = "2021"
+description = "Rig framework Tool implementations for coding tools"
+repository = "https://github.com/Sewer56/rig-coding-tools"
+license = "Apache-2.0"
+include = ["src/**/*"]
+readme = "README.md"
+
+[dependencies]
+async-trait = "0.1"
+coding-tools-core = { version = "0.1.0", path = "../coding-tools-core" }
+reqwest = { version = "0.12", default-features = false, features = ["rustls-tls"] }
+rig-core = { version = "0.27", default-features = false, features = ["reqwest-rustls"] }
+schemars = "1.2.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tokio = { version = "1.0", features = ["fs", "process", "time", "io-util", "sync"] }
+
+[dev-dependencies]
+tempfile = "3.10"
+tokio = { version = "1.0", features = ["rt", "macros"] }
