+++ src/coding-tools-rig/src/allowed/grep.rs
@@ -1,0 +1,193 @@
+//! Grep content search tool using AllowedPathResolver.
+
+use coding_tools_core::operations::grep_search;
+use coding_tools_core::path::AllowedPathResolver;
+use coding_tools_core::{ToolError, ToolOutput, ToolResult};
+use rig::completion::ToolDefinition;
+use rig::tool::Tool;
+use schemars::{schema_for, JsonSchema};
+use serde::Deserialize;
+use std::fmt::Write;
+use std::path::{Path, PathBuf};
+
+const DEFAULT_LIMIT: usize = 100;
+const MAX_LIMIT: usize = 2000;
+const MAX_LINE_LENGTH: usize = 2000;
+
+fn default_limit() -> Option<usize> {
+    Some(DEFAULT_LIMIT)
+}
+
+/// Arguments for the grep tool.
+#[derive(Debug, Deserialize, JsonSchema)]
+pub struct GrepArgs {
+    /// Regex pattern to search for in file contents.
+    pub pattern: String,
+    /// Relative directory path to search in (within allowed directories).
+    pub path: String,
+    /// Optional file glob filter (e.g., "*.rs", "*.{ts,tsx}").
+    #[serde(default)]
+    pub include: Option<String>,
+    /// Maximum number of matches to return (default: 100, max: 2000).
+    #[serde(default = "default_limit")]
+    pub limit: Option<usize>,
+}
+
+/// Tool for searching file contents within allowed directories.
+#[derive(Debug, Clone)]
+pub struct GrepTool<const LINE_NUMBERS: bool = true> {
+    resolver: AllowedPathResolver,
+}
+
+impl<const LINE_NUMBERS: bool> GrepTool<LINE_NUMBERS> {
+    /// Creates a new grep tool restricted to the given directories.
+    pub fn new(allowed_paths: impl IntoIterator<Item = impl AsRef<Path>>) -> ToolResult<Self> {
+        let paths: Vec<PathBuf> = allowed_paths
+            .into_iter()
+            .map(|p| p.as_ref().to_path_buf())
+            .collect();
+        Ok(Self {
+            resolver: AllowedPathResolver::new(paths)?,
+        })
+    }
+
+    /// Creates a new grep tool with an existing resolver.
+    pub fn with_resolver(resolver: AllowedPathResolver) -> Self {
+        Self { resolver }
+    }
+}
+
+impl<const LINE_NUMBERS: bool> Tool for GrepTool<LINE_NUMBERS> {
+    const NAME: &'static str = "grep";
+
+    type Error = ToolError;
+    type Args = GrepArgs;
+    type Output = ToolOutput;
+
+    async fn definition(&self, _prompt: String) -> ToolDefinition {
+        ToolDefinition {
+            name: Self::NAME.to_string(),
+            description: "Search file contents using regex patterns within allowed directories. \
+                          Paths are relative to configured base directories."
+                .to_string(),
+            parameters: serde_json::to_value(schema_for!(GrepArgs))
+                .expect("schema serialization should not fail"),
+        }
+    }
+
+    async fn call(&self, args: Self::Args) -> Result<Self::Output, Self::Error> {
+        let pattern = args.pattern.trim();
+        if pattern.is_empty() {
+            return Err(ToolError::InvalidPattern(
+                "pattern must not be empty".into(),
+            ));
+        }
+
+        let limit = args.limit.unwrap_or(DEFAULT_LIMIT).min(MAX_LIMIT);
+        if limit == 0 {
+            return Err(ToolError::InvalidPattern(
+                "limit must be greater than zero".into(),
+            ));
+        }
+
+        let include = args.include.as_deref().and_then(|s| {
+            let trimmed = s.trim();
+            if trimmed.is_empty() {
+                None
+            } else {
+                Some(trimmed)
+            }
+        });
+
+        let result = grep_search(&self.resolver, pattern, include, &args.path, limit)?;
+
+        if result.files.is_empty() {
+            return Ok(ToolOutput::new("No matches found."));
+        }
+
+        // Format output grouped by file
+        let mut output = String::with_capacity(4096);
+        let _ = writeln!(&mut output, "Found {} matches", result.match_count);
+
+        for file in &result.files {
+            let _ = writeln!(&mut output, "\n{}:", file.path);
+            for m in &file.matches {
+                let truncated_text = if m.line_text.len() > MAX_LINE_LENGTH {
+                    &m.line_text[..MAX_LINE_LENGTH]
+                } else {
+                    &m.line_text
+                };
+                if LINE_NUMBERS {
+                    let _ = writeln!(&mut output, "  L{}: {}", m.line_num, truncated_text);
+                } else {
+                    let _ = writeln!(&mut output, "  {}", truncated_text);
+                }
+            }
+        }
+
+        if result.truncated {
+            let _ = write!(&mut output, "\n(Results truncated at {} matches)", limit);
+        }
+
+        Ok(if result.truncated {
+            ToolOutput::truncated(output)
+        } else {
+            ToolOutput::new(output)
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use tempfile::TempDir;
+
+    #[tokio::test]
+    async fn finds_matching_content() {
+        let dir = TempDir::new().unwrap();
+        std::fs::write(dir.path().join("test.txt"), "hello world").unwrap();
+
+        let tool: GrepTool<true> = GrepTool::new([dir.path()]).unwrap();
+        let result = tool
+            .call(GrepArgs {
+                pattern: "hello".to_string(),
+                path: ".".to_string(),
+                include: None,
+                limit: None,
+            })
+            .await
+            .unwrap();
+        assert!(result.content.contains("Found 1 matches"));
+        assert!(result.content.contains("L1: hello world"));
+    }
+
+    #[tokio::test]
+    async fn rejects_path_traversal() {
+        let dir = TempDir::new().unwrap();
+        let tool: GrepTool = GrepTool::new([dir.path()]).unwrap();
+        let result = tool
+            .call(GrepArgs {
+                pattern: "test".to_string(),
+                path: "../../../etc".to_string(),
+                include: None,
+                limit: None,
+            })
+            .await;
+        assert!(matches!(result, Err(ToolError::InvalidPath(_))));
+    }
+
+    #[tokio::test]
+    async fn rejects_empty_pattern() {
+        let dir = TempDir::new().unwrap();
+        let tool: GrepTool = GrepTool::new([dir.path()]).unwrap();
+        let result = tool
+            .call(GrepArgs {
+                pattern: "   ".to_string(),
+                path: ".".to_string(),
+                include: None,
+                limit: None,
+            })
+            .await;
+        assert!(matches!(result, Err(ToolError::InvalidPattern(_))));
+    }
+}
