--- src/rig-coding-tools/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "rig-coding-tools"
-version = "0.1.0"
-edition = "2021"
-description = "Basic coding tools for rig based LLM agents"
-repository = "https://github.com/Sewer56/rig-coding-tools"
-license = "MIT"
-include = ["src/**/*"]
-readme = "README.MD"
-
-[dependencies]
-async-trait = "0.1"
-rig-core = { version = "0.27", default-features = false, features = ["reqwest-rustls"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-schemars = "1.2.0"
-thiserror = "2.0"
-tokio = { version = "1", features = ["fs", "process", "time", "io-util"] }
-glob = "0.3"
-memchr = "2.7"
-ignore = "0.4"
-regex = "1.11"
-grep = "0.4"
-reqwest = { version = "0.12", default-features = false, features = ["rustls-tls"] }
-html-to-markdown-rs = "2.16"
-
-[dev-dependencies]
-tokio = { version = "1", features = ["full", "test-util"] }
-tempfile = "3"
-wiremock = "0.6"
-anyhow = "1.0"
