--- src/rig-coding-tools/README.MD
@@ -1,58 +1,0 @@
-# rig-coding-tools
-
-[![Crates.io](https://img.shields.io/crates/v/rig-coding-tools.svg)](https://crates.io/crates/rig-coding-tools)
-[![Docs.rs](https://docs.rs/rig-coding-tools/badge.svg)](https://docs.rs/rig-coding-tools)
-[![CI](https://github.com/Sewer56/rig-coding-tools/actions/workflows/rust.yml/badge.svg)](https://github.com/Sewer56/rig-coding-tools/actions)
-
-<!-- 
-    This README is displayed on crates.io & docs.rs
-    Write detailed documentation here for crate users.
--->
-
-Basic coding tools for rig based LLM agents
-
-## Features
-
-<!-- TODO: Replace with actual features of your crate -->
-
-- Feature 1
-- Feature 2
-- Feature 3
-
-## Installation
-
-Add this to your `Cargo.toml`:
-
-```toml
-[dependencies]
-rig-coding-tools = "0.1.0"
-```
-### Feature Flags
-
-| Feature | Description |
-| ------- | ----------- |
-| `std` | Enable standard library support (disabled by default for `no_std` compatibility) |
-
-## Usage
-
-<!-- TODO: Add comprehensive examples showing common use cases -->
-
-### Basic Example
-
-```rust
-use rig_coding_tools::{ToolError, ToolResult};
-
-// Tool implementations will be added in future releases
-fn example() -> ToolResult<()> {
-    // Tools will follow the rig-core Tool trait pattern
-    Ok(())
-}
-```
-
-### Advanced Example
-
-Examples will be added as tools are implemented.
-
-## License
-
-Licensed under Apache 2.0
