--- src/rig-coding-tools/src/tools/mod.rs
@@ -1,22 +1,0 @@
-//! Tool implementations for rig-based LLM agents.
-//!
-//! Each submodule implements a specific tool following the `rig_core::tool::Tool` trait.
-
-// Tool submodules will be added here as they are implemented:
-pub mod bash;
-pub mod edit;
-pub mod glob;
-pub mod grep;
-pub mod read;
-pub mod task;
-pub mod todo;
-pub mod webfetch;
-pub mod write;
-// pub mod skill;
-
-// Re-exports
-pub use bash::BashTool;
-pub use grep::{GrepArgs, GrepFileMatches, GrepLineMatch, GrepOutput, GrepTool};
-pub use read::ReadTool;
-pub use todo::{Todo, TodoPriority, TodoReadTool, TodoState, TodoStatus, TodoWriteTool};
-pub use webfetch::WebFetchTool;
